Revert "Revert "Tentativo per vedere se su VM funge""
This reverts commit fefaf1079811b0264468ca2e1d75946db5f8c452.

+++ src/Graph/Graph.h
@@ -50,7 +50,7 @@ public:
     int num_of_edges() { return sizeE; }
     void setSizeNodes(int value) { sizeN = value; }
     void setSizeEdges(int value) { sizeE = value; }
-    int getNodeWeight(int nodeIndex) const {
+    float getNodeWeight(int nodeIndex) const {
         return Nodes.at(nodeIndex).weight;
     }
     double getNodeWeightAvg() const {
